+++ src/com/dci/intellij/dbn/database/mysql/MySqlMetadataInterface.java
@@ -21,7 +21,7 @@ public class MySqlMetadataInterface extends DatabaseMetadataInterfaceImpl {
     }
 
     @Override
-    public ResultSet loadMethodArguments(String ownerName, String methodName, String methodType, int overload, DBNConnection connection) throws SQLException {
+    public ResultSet loadMethodArguments(String ownerName, String methodName, String methodType, short overload, DBNConnection connection) throws SQLException {
         try {
             return super.loadMethodArguments(ownerName, methodName, methodType, overload, connection);
         } catch (SQLException e) {
