+++ src/com/dci/intellij/dbn/object/common/list/DBObjectList.java
@@ -38,6 +38,6 @@ public interface DBObjectList<T extends DBObject> extends BrowserTreeNode, Dynam
     List<T> getObjects();
     List<T> getObjects(String name);
     T getObject(String name);
-    T getObject(String name, int overload);
+    T getObject(String name, short overload);
     T getObject(String name, String parentName);
 }
